requirements.txt
sqlalchemy
psycopg2
discord.py
requests
python-dotenv
beautifulsoup4